feat(consoleui): add new command to show main menu
Added a new command `ShowMainMenuCommand` that displays the main menu options and allows the user to navigate through them using arrow keys and select an option by pressing Enter. The command uses a `SinglePanelViewModel` to display the menu entries and a `SelectionService` for user interaction. The chosen option is used to create and execute a submenu command.